requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.41
psycopg2==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.11.5
pydantic-settings==2.10.0
boto3==1.38.33
aiosqlite==0.21.0
websockets==14.2
asyncio-mqtt==0.16.2
python-dotenv==1.1.1
pyotp
kiteconnect
pymysql